Experiment with binding the implementation of only one module
In the previous commit, all parsed modules were run through all
binding phases including the BindImplementation phase.  This commit
is an experiment with only running the final binding phase, which
is the slowest, on the module corresponding to the file currently
being edited.

Performance seems to improve quite a bit.  There is still a major
memory leak somewhere but at least MonoDevelop is now usable
beyond 5-10 minutes.

This build will compile and run but is somewhat unstable.